Resolved circular modular design
