+++ enginecore/enginecore/state/api/__init__.py
@@ -4,3 +4,4 @@ from enginecore.state.api.pdu import IPDUStateManager
 from enginecore.state.api.outlet import IOutletStateManager
 from enginecore.state.api.static import IStaticDeviceManager
 from enginecore.state.api.server import IServerStateManager, IBMCServerStateManager
+from enginecore.state.api.environment import ISystemEnvironment
